Refactor application: Move Session and SecurityService to security package Refactore html pages
