+++ src/main/resources/editeditem.html
@@ -1,17 +1,11 @@
-    <!DOCTYPE html>
-    <html xmlns="http://www.w3.org/1999/xhtml">
-    <head>
-        <!-- Required meta tags -->
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-
-        <!-- Bootstrap CSS -->
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-
-        <title>Edited item</title>
-    </head>
-    <body>
-    <div class="container">
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head>
+    <#include "templates/parts/header.html">
+    <title>Edited item</title>
+</head>
+<body>
+<div class="container">
     <h1 class="text-center">Item edited in DB</h1>
 
     <p class="text-left">Item updated values:</p>
@@ -20,12 +14,12 @@
     <p>creationdate: ${creationdate}</p>
     <p>description: ${description}</p>
 
-    <form action="/items" class="inline">
-        <button class="float-left submit-button" >Show all items</button>
-    </form>
+    <div class="float-left">
+        <a href="/items" class="submit-button">
+            Show all items
+        </a>
     </div>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa"
-            crossorigin="anonymous"></script>
-    </body>
-    </html>
+</div>
+<#include "templates/parts/footer.html">
+</body>
+</html>
