+++ src/main/resources/edititem.html
@@ -1,21 +1,17 @@
-    <!DOCTYPE html>
-    <html lang="en">
-    <head>
-        <!-- Required meta tags -->
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-
-        <!-- Bootstrap CSS -->
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-
-        <title>Edit item</title>
-    </head>
-    <body>
-    <div class="container">
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <#include "templates/parts/header.html">
+    <title>Edit item</title>
+</head>
+<body>
+<div class="container">
     <h1 class="text-center">EDIT ITEM</h1>
     <p class="text-left">Please put values into the form you want to update:</p>
 
     <form action="/edititem" method="POST">
+        <input type="hidden" name="id" value=${id}>
+
         <label for="name">Name:</label><br>
         <input type="text" pattern="[A-Za-z0-9 ]{3,}" id="name" name="name" value="${name}"><br>
         <label for="price">Price:</label><br>
@@ -24,12 +20,10 @@
         <input type="datetime-local" id="daytime" name="creationday" value="${creationdate}"><br>
         <label for="description">Description:</label><br>
         <input type="text" id="description" name="description" value="${description}"><br>
+
         <input type="submit" value="Update">
-        <input type="hidden" name="id" value=${id}>
     </form>
-    </div>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa"
-            crossorigin="anonymous"></script>
-    </body>
-    </html>
+</div>
+<#include "templates/parts/footer.html">
+</body>
+</html>
