+++ src/main/resources/failedlogin.html
@@ -8,10 +8,10 @@
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
           integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-
     <title>login failed</title>
 </head>
 <body>
+
 <div class="container">
     <p class="text-left">Login failed. Incorrect password or user does not exist.</p>
     <p class="text-left">Please try again or get back to items list</p>
